Containerize the project add User MVC
**Description**

This update adds the user model, view and controller. For the Record ID
I opted for a UUID because this is far more secure than simply
incrementing a numerical ID. This was a little tricky to do in SQLite,
since I have normally used a postgres databse for recent rails projects.

I opted to use sqllite for simplicity and ease of containerization
limiting the need for a second container in the docker config.

I user ChatGPT to generate the SCSS to speed things along since I am a
little rusty on CSS. Otherwise, the routes are fairly straight forward.
To validate in a dev environment, I also created a quick db seed file to
generate a couple of records since the proposal indicated that users
would already exist and we would be updating the user password value.

This project also uses two custom validators that I have built in a
previous project.