Use bcrypt has_secure_password
**Description**

This update adds a sessions controller for basic authentication and
leverages the bcrypt gem for secure password validations.

I chose to add the password validation as a second migrations to keep
record changes encapsulated. Had I added it in the beginning I would
have had to tackle to many changes at once.

I included a simplistic sessions controller to allow creating and
signing in as a user. There is no RBAC applied at this time but that
would likely be the very next step in hardening this app.