Fixing Password update method
**Description**

Things were simpler if I provided a current and password value instead
of a password and new_password value. This allows the validations to run
without much trickery. I updated the specs for the model and stubbed out
some specs for the controller. Due to time constraints I did not opt to
add the controller specs in full. However, I wanted to indicated that I
was thinking about them.

At this point the code meets the minimum requirements of the challenge.
It is worth noting that unlike the original sequence diagram this does
not implement any client side validations/rendering. In an effort to
keep this project light I opted to simply implement server-side
rendering. However, in a more modern implementation I would likely use
json validator with
[stimulus](https://www.mikewilson.dev/posts/client-side-form-validation-with-stimulus-and-rails/) to have some client side form validations
over pushing everything to the server.